quantity value setter updated to handle decimals and updated expetected test data where required

+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/mapper/QuantityMapper.java
@@ -69,6 +69,6 @@ public class QuantityMapper {
 
     private void setQuantityValueAndUnit(Quantity quantity, String value, String unit, List<PQR> translation) {
         setUnit(quantity, unit, translation);
-        quantity.setValue((long) Double.parseDouble(value));
+        quantity.setValue(Double.parseDouble(value));
     }
 }
